requirements.txt
exceptiongroup==1.2.0
greenlet==3.0.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
PyMySQL==1.1.0
pytest==7.4.3
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.8.0
